Remove src_addr from EXT-ROUTER
We've found that this attribute contains a very wide range IPs, which doesn't seem to be what you'd expect for a router.  E.g. we're seeing the entire space of IPs in various ranges, like 10.1.1.0/24 and others for the same entity.  Our suspicion is that this value is the src_addr of TCP flows / devices that the router may be talking to / have entries in its MAC table, etc.

This is problematic for us as the tag will have a lot of churn (e.g. we're seeing millions of changes per minute for an account with just 250 routers) and cause excessive writes to the entity store (which works under the assumption of storing objects that don't change that often).

We're removing this for now so that we can re-enable the entity type.