EXT-SERVICE: match on service.name
Restoring an original rule that matched based on the pressence of the
attribute.  Added a warning to note that this rule matches too broadly
in some instrumentation sources, and we're ignoring in certain ingestion
paths until we add a configuration in rules that allows selecting the
right ingest paths to apply it.

Signed-off-by: Galo Navarro <gnavarro@newrelic.com>