[infra-kubernetes_pod] Use k8s.pod.isScheduled for pods
The synthesis rule for pods uses the k8s.pod.isReady metric. This metric can
be null if the Pod is never scheduled; Pods that can't be scheduled and
are stuck in Pending do not report the metric and are not being indexed.

We are changing the rule to use the k8s.pod.isScheduled metric since
it's reported by Pods even if they are not scheduled.