Recover host systhesize process from SAP data source. 	modified:   ../definition.yml 	new file:   SAPHost.json
It was decided to use infra-host entity for host information genrated by SAP datasource. However, some late changes on host definition broke the process as it check on a data souce attibute which are not available for SAP generated metrics. This change is intend to recover the support of sythesizing host entity for data coming from SAP data source.