fix: Adjust the failures golden metric for synthetic monitors.
ticket: https://issues.newrelic.com/browse/NR-112045
context:
The current synthetic golden metric configuration for failures is causing alert configuration issues. Since the WHERE condition limits this to failures only, the entity status is not getting set to green when the monitor passes. It is either gray (no data) or red. Alert configurations using the new condition builder flow would also not close correctly as it would require a signal loss configuration + and a timer.

This new configuration will allow all check results to be passed to the alerts pipeline, which will allow all default alert settings to work properly and for the entity state to be updated correctly.