Add a generic rule for APM->DB relationships
This new rule allows any hostname as DB endpoint, not limiting it to RDS.
It includes a more strict regex which only includes those providers that are supported in the `DATABASE` candidate category.
The regex also filters out those unsupported metric names that contain a `;` in the endpoint (legacy format to concat different endpoints in a single metric).